+++ internal/cmd/rootcmd/rootcmd.go
@@ -21,6 +21,9 @@ func Execute() {
 				return Run(cmd.Context(), &cfg)
 			},
 			SilenceUsage: true,
+			// We can silence the errors because cobra.CheckErr below will print
+			// the returned error and set the exit code to 1.
+			SilenceErrors: true,
 		}
 		versionCmd = &cobra.Command{
 			Use:   "version",